Add tests for battery level
