+++ src/app/components/Battery/Battery.js
@@ -1,10 +1,8 @@
 import React, { Component } from "react"
-import NumericValue from "./NumericValue"
-import MqttSubscriptions from "../mqtt/MqttSubscriptions"
-import MqttTopicWildcard from "../mqtt/MqttTopicWildcard"
-import PropTypes from "prop-types"
-
-import { BATTERY_STATE } from "../utils/constants"
+import NumericValue from "./../NumericValue"
+import MqttSubscriptions from "../../mqtt/MqttSubscriptions"
+import MqttTopicWildcard from "../../mqtt/MqttTopicWildcard"
+import BatteryLevel from "./BatteryLevel"
 
 const secondaryBatteriesFeatureEnabled = true
 
@@ -17,32 +15,6 @@ const getTopics = portalId => {
     power: `N/${portalId}/system/0/Dc/Battery/Power`,
     voltage: `N/${portalId}/system/0/Dc/Battery/Voltage`,
     mainBattery: `N/${portalId}/system/0/ActiveBatteryService`
-  }
-}
-
-function batteryStateFormatter(value) {
-  switch (value) {
-    case BATTERY_STATE.CHARGING:
-      return "Charging"
-    case BATTERY_STATE.DISCHARGING:
-      return "Discharging"
-    case BATTERY_STATE.IDLE:
-      return "Idle"
-  }
-}
-function batteryTimeToGoFormatter(timeToGo) {
-  const secs = parseInt(timeToGo)
-  if (!isNaN(secs)) {
-    const days = Math.floor(secs / 86400)
-    const hours = Math.floor((secs - days * 86400) / 3600)
-    const minutes = Math.floor((secs - hours * 3600) / 60)
-    const seconds = Math.floor(secs - minutes * 60)
-    if (days) return `${days}d ${hours}h`
-    else if (hours) return `${hours}h ${minutes}m`
-    else if (minutes) return `${minutes}m ${seconds}s`
-    else return `${seconds}s`
-  } else {
-    return null
   }
 }
 
@@ -84,36 +56,11 @@ const secondaryBatteries = (instances, mainBatteryInstance, portalId) => {
   })
 }
 
-const BatteryLevelContainer = props => {
-  const batteryStateLabel = batteryStateFormatter(props.state)
-  const timeToGoLabel = batteryTimeToGoFormatter(props.timeToGo)
-  const showTimetoGo = props.state === BATTERY_STATE.DISCHARGING
-  return (
-    <div className={"metric__battery-level-container" + (showTimetoGo ? " metric__battery-level-container--col" : "")}>
-      <div className="text--bottom-align">
-        <span className="text text--bold text--large">{props.soc ? props.soc : ""}</span>
-        <span className="text text--small metric__battery-state">
-          {props.soc ? "%" : ""}
-          &nbsp;
-          {batteryStateLabel || ""}
-        </span>
-      </div>
-      {showTimetoGo && props.timeToGo ? <p className="text text--small">{timeToGoLabel}</p> : ""}
-    </div>
-  )
-}
-
-BatteryLevelContainer.propTypes = {
-  state: PropTypes.oneOf(Object.values(BATTERY_STATE)),
-  soc: PropTypes.number,
-  timeToGo: PropTypes.number
-}
-
 const Battery = props => {
   return (
     <div className="metric metric__container metric__battery">
       <div className="metric__container--left">
-        <img src={require("../../images/icons/battery.svg")} className="metric__icon" />
+        <img src={require("../../../images/icons/battery.svg")} className="metric__icon" />
         <div className="metric__value-container">
           <p className="text text--medium">Battery</p>
           <div className="metric__values">
@@ -126,7 +73,7 @@ const Battery = props => {
             secondaryBatteries(props.batteryInstances, props.mainBattery, props.portalId)}
         </div>
       </div>
-      <BatteryLevelContainer state={props.state} soc={props.soc} timeToGo={props.timeToGo} />
+      <BatteryLevel state={props.state} soc={props.soc} timeToGo={props.timeToGo} />
     </div>
   )
 }
